requirements.txt
django==1.11.3
django-model-utils==3.0.0
django-annoying==0.10.3
django-attachments==1.1.1
django-material==1.0.0
# psycopg2==2.7.1